+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Basic Qt5 theming support using installed system themes
+- Allow for a pext/config.ini file to overwrite some default configuration
 
 ### Changed
 - Get list of installable modules from pext.hackerchick.me instead of pext.github.io